requirements.txt
grpcio
protobuf
torch
torchvision